// swiyu-ecosystem-portal-ui/src/app/api/generated/model/trust-onboarding-submission-request.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Address} from './address';
import {Contact} from './contact';
import {MultiLanguageText} from './multi-language-text';

export interface TrustOnboardingSubmissionRequest {
  partnerId?: string;
  entityName?: MultiLanguageText;
  entityAddress?: Address;
  entityEmail?: string;
  contactPerson?: Contact;
  registryIds?: {[key: string]: string};
  isRegisteredInCommercialRegister?: boolean;
  correspondingLanguage?: TrustOnboardingSubmissionRequest.CorrespondingLanguageEnum;
  dids?: Array<string>;
  requestedPartnerType?: TrustOnboardingSubmissionRequest.RequestedPartnerTypeEnum;
}
export namespace TrustOnboardingSubmissionRequest {
  export const CorrespondingLanguageEnum = {
    En: 'EN',
    De: 'DE',
    Fr: 'FR',
    It: 'IT',
    Rm: 'RM'
  } as const;
  export type CorrespondingLanguageEnum = (typeof CorrespondingLanguageEnum)[keyof typeof CorrespondingLanguageEnum];
  export const RequestedPartnerTypeEnum = {
    GovernmentalInstitution: 'GOVERNMENTAL_INSTITUTION',
    Business: 'BUSINESS',
    Individual: 'INDIVIDUAL',
    Unknown: 'UNKNOWN'
  } as const;
  export type RequestedPartnerTypeEnum = (typeof RequestedPartnerTypeEnum)[keyof typeof RequestedPartnerTypeEnum];
}